requirements.txt
python-dotenv
groq
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
pymupdf4llm
pydantic
pydantic-settings
numpy
gradio
plotly
pandas
google-generativeai
